+++ server/src-lib/Hasura/RQL/DDL/Metadata.hs
@@ -367,7 +367,12 @@ runReloadMetadata (ReloadMetadata reloadRemoteSchemas reloadSources reloadRecrea
           }
 
   buildSchemaCacheWithOptions (CatalogUpdate $ Just recreateEventTriggersSources) cacheInvalidations metadata
-  pure successMsg
+  inconsObjs <- scInconsistentObjs <$> askSchemaCache
+  pure . encJFromJValue . J.object $
+    [ ("message" :: Text) J..= ("success" :: Text),
+      "is_consistent" J..= null inconsObjs
+    ]
+      <> ["inconsistent_objects" J..= inconsObjs | not (null inconsObjs)]
 
 runDumpInternalState ::
   (QErrM m, CacheRM m) =>